fish: get rid of `test_{source,exec}`
I was having the same issue as [this
person](https://github.com/asdf-vm/asdf-nodejs/issues/253), but the
solution in that thread didn't solve it for me. I found that for some
reason, doing the sourcing inside of `test_source` as opposed to inline
was the culprit, so we're just ditching those two functions to avoid any
future wonkiness.